requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
python-multipart>=0.0.12
python-dotenv>=1.0.1
opencv-python-headless>=4.10.0
Pillow>=10.4.0
numpy>=2.2.0
requests>=2.32.0
email-validator>=2.2.0
psycopg2-binary>=2.9.9
cloudinary>=1.40.0
